GH-175: make it possible to supply collections w/o flat-mapping
It was impossible to supply a collection as is into test method. It was automatically expanded. When we don't use `flatMap`, collections should always be injected in the original form w/o elements' extraction. This update adjust the collections processing behavior.

Apart from that, it covers dependency and template updates.